CORE-10453 : 1. Removed AliasIdentityDataService.kt and HardcodedAliasIdentityServiceImpl.kt 2. Created HardcodedAliasIdentityHelper 3. cpiName changed to hostNetwork
