--- components/interop/interop-service/src/main/kotlin/net/corda/interop/service/AliasIdentityDataService.kt
@@ -1,7 +1,0 @@
-package net.corda.interop.service
-
-import net.corda.v5.interop.InteropGroupInfo
-
-interface AliasIdentityDataService {
-    fun getAliasIdentityData() : List<InteropGroupInfo>
-}
